Move pkg/internal/tbtctest to pkg/tbtc/internal/test
For consistency with changes proposed in https://github.com/keep-network/keep-core/pull/3626#discussion_r1227884983
we move the tbtctest package as it's used internally in pkg/tbtc only.
