+++ pkg/tbtc/deposit_sweep_test.go
@@ -14,13 +14,13 @@ import (
 	"github.com/keep-network/keep-core/pkg/tecdsa"
 
 	"github.com/keep-network/keep-core/pkg/bitcoin"
-	"github.com/keep-network/keep-core/pkg/internal/tbtctest"
 	"github.com/keep-network/keep-core/pkg/internal/testutils"
+	"github.com/keep-network/keep-core/pkg/tbtc/internal/test"
 )
 
 // TODO: Think about covering unhappy paths for specific steps of the deposit sweep action.
 func TestDepositSweepAction_Execute(t *testing.T) {
-	scenarios, err := tbtctest.LoadDepositSweepTestScenarios()
+	scenarios, err := test.LoadDepositSweepTestScenarios()
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -228,7 +228,7 @@ func TestDepositSweepAction_Execute(t *testing.T) {
 }
 
 func TestAssembleDepositSweepTransaction(t *testing.T) {
-	scenarios, err := tbtctest.LoadDepositSweepTestScenarios()
+	scenarios, err := test.LoadDepositSweepTestScenarios()
 	if err != nil {
 		t.Fatal(err)
 	}
