+++ pkg/tbtc/internal/test/marshaling.go
@@ -1,4 +1,4 @@
-package tbtctest
+package test
 
 import (
 	"crypto/ecdsa"
