+++ pkg/tbtc/internal/test/testdata/deposit_sweep_scenario_0.json
