+++ pkg/tbtc/internal/test/testdata/deposit_sweep_scenario_2.json
