+++ pkg/tbtc/internal/test/testdata/deposit_sweep_scenario_3.json
