+++ pkg/tbtc/internal/test/testdata/redemption_scenario_0.json
