+++ pkg/tbtc/internal/test/testdata/redemption_scenario_1.json
