+++ pkg/tbtc/internal/test/testdata/redemption_scenario_2.json
