+++ pkg/tbtc/internal/test/testdata/redemption_scenario_3.json
