+++ pkg/tbtc/internal/test/testdata/redemption_scenario_4.json
