+++ pkg/tbtc/internal/test/testdata/redemption_scenario_5.json
