+++ pkg/tbtc/redemption_test.go
@@ -1,16 +1,17 @@
 package tbtc
 
 import (
-	"github.com/go-test/deep"
 	"testing"
 
+	"github.com/go-test/deep"
+
 	"github.com/keep-network/keep-core/pkg/bitcoin"
-	"github.com/keep-network/keep-core/pkg/internal/tbtctest"
 	"github.com/keep-network/keep-core/pkg/internal/testutils"
+	"github.com/keep-network/keep-core/pkg/tbtc/internal/test"
 )
 
 func TestAssembleRedemptionTransaction(t *testing.T) {
-	scenarios, err := tbtctest.LoadRedemptionTestScenarios()
+	scenarios, err := test.LoadRedemptionTestScenarios()
 	if err != nil {
 		t.Fatal(err)
 	}
